Implemented abstract labels based on i18n for custom tags
A LabelMaker interface is introduced. The example below generates the label (name) for the 'author' attribute of the 'quote' custom tag.
```
$labelMaker->makeLabel('label', ['quote', 'author']);
```

It comes with a TranslationBasedLabelMaker that uses the Translator. Configured patterns mapped to label identifiers (ex: 'label', 'description')
are used to generate a translation key, passed to the Translator. If no translation is set, it may generate a default label based on the
last item.

The TranslationBasedLabelMaker is instanciated as multiple services, one for each label type (ex: custom tag, custom tag attribute...).