Multilingual support for ezbinaryfile preview
For now, in the admin, if you display the main view of a content in another language than the default one, it will display the correct link (with the filename in the correct language) but if you click on the download button, you will always get the content of the file in the language of the admin.

I hope that this fix could be convenient, I tried it on two of my projects and it fixed the wrong behaviour with no side effect so far. I am using native eZ elements (the prioritizedFieldLanguageCode of the content and the preconfigured inLanguage option for the route).