Set config providers which use dynamic settings to lazy
This fixes an issue where the dynamic parameters in arguments are injected to the service in the `default` scope instead of `admin` scope, resulting in admin content edit interface without any styles.

I have not been able to deduce why this is happening, though, since it seems that clean eZ Platform install works fine. It maybe a combination of some other config, but anyway, setting these services to lazy helps.